requirements.txt
PILLOW
Pytesseract
streamlit